Финальная доработка. Основное: Документирование кода, созданы интерфейсы, оформлен файл REDMI.
