+++ README.md
@@ -1,1 +1,45 @@
-Материалы для выполнения курсовой работы учениками профессии java-разработчик. 
+**Дипломный проект по профессии java developer.**
+
+***Онлайн университет SkyPro, курс java developer.***
+
+Проект выполнил Шадрин Алексей.
+
+**Цель проекта:** закрепление полученных в ходе обучения знаний и приобретение практических навыков.
+
+**Задачи проекта:**
+
+Написать серверную часть сайта по продаже вещей включающую в себя:
+- Авторизацию и аутентификацию пользователей;
+- Распределение ролей между пользователями: пользователь и администратор;
+- CRUD для объявлений на сайте;
+- Возможность оставлять отзывы под объявлениями;
+- Показывать и сохранять картинки объявлений.
+
+**Исходные данные:**
+- Спецификация openapi.yaml файл;
+- Frontend часть, с возможностью запуска через Docker.
+
+**При выполнении проекта был реализован следующий функционал:**
+1. Без авторизации:
+- Просмотр списка объявлений;
+- Регистрация нового пользователя;
+
+2. Авторизация с ролью "USER":
+- Доступ к профилю и его редактирование;
+- Смена пароля;
+- Создание новых объявлений;
+- Редактирование и удаление своих объявлений;
+- Возможность оставить комментарии под объявлениями;
+- Редактирование и удаление своих комментариев;
+- Редактирование изображений в своих объявлениях;
+
+3. Авторизация с ролью "ADMIN", к вышеперечисленному добавляются возможности:
+- Редактирование и удаление любых объявлений;
+- Редактирование и удаление любых комментариев;
+- Редактирование изображений в любых объявлениях;
+
+**Для выполнения задач задействовано:**
+- Java language;
+- Spring Framework, Spring Security;
+- Lombok, Liquibase, Mapstruct, Springdoc;
+- Database PostgreSQL.
