+++ src/main/java/ru/skypro/homework/controller/AuthController.java
@@ -1,8 +1,13 @@
 package ru.skypro.homework.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,6 +28,19 @@ public class AuthController {
 
     private final AuthService authService;
 
+    @Operation(summary = "login",
+            tags = "Авторизация",
+            requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = LoginReq.class))
+            ),
+            responses = {
+                    @ApiResponse(responseCode = "200"),
+                    @ApiResponse(responseCode = "401", content = @Content),
+                    @ApiResponse(responseCode = "403", content = @Content),
+                    @ApiResponse(responseCode = "404", content = @Content),
+            }
+    )
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody LoginReq req) {
         if (authService.login(req.getUsername(), req.getPassword())) {
@@ -32,6 +50,19 @@ public class AuthController {
         }
     }
 
+    @Operation(summary = "register",
+            tags = "Авторизация",
+            requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = RegisterReq.class))
+            ),
+            responses = {
+                    @ApiResponse(responseCode = "201", content = @Content),
+                    @ApiResponse(responseCode = "401", content = @Content),
+                    @ApiResponse(responseCode = "403", content = @Content),
+                    @ApiResponse(responseCode = "404", content = @Content),
+            }
+    )
     @PostMapping("/register")
     public ResponseEntity<?> register(@RequestBody RegisterReq req) {
         Role role = req.getRole() == null ? USER : req.getRole();
