+++ src/main/java/ru/skypro/homework/service/AuthService.java
@@ -5,6 +5,7 @@ import ru.skypro.homework.dto.RegisterReq;
 import ru.skypro.homework.dto.Role;
 
 public interface AuthService {
+
     boolean login(String userName, String password);
     boolean register(RegisterReq registerReq, Role role);
     void checkAccess(String userNameAuthor, Authentication authentication);
